chore: move wrapped signal impl to separate file
