+++ packages/qwik/src/core/signal/impl/wrapped-signal-impl.ts
@@ -1,0 +1,96 @@
+import { assertFalse } from '../../shared/error/assert';
+import { QError, qError } from '../../shared/error/error';
+import type { Container, HostElement } from '../../shared/types';
+import { ChoreType } from '../../shared/util-chore-type';
+import { trackSignal } from '../../use/use-core';
+import { _EFFECT_BACK_REF, NEEDS_COMPUTATION } from '../flags';
+import { triggerEffects } from '../signal';
+import type { BackRef } from '../signal-cleanup';
+import type { AllSignalFlags, EffectSubscription } from '../types';
+import { EffectProperty, SignalFlags, WrappedSignalFlags } from '../types';
+import { SignalImpl } from './signal-impl';
+
+export class WrappedSignalImpl<T> extends SignalImpl<T> implements BackRef {
+  $args$: any[];
+  $func$: (...args: any[]) => T;
+  $funcStr$: string | null;
+
+  $flags$: AllSignalFlags;
+  $hostElement$: HostElement | null = null;
+  $forceRunEffects$: boolean = false;
+  [_EFFECT_BACK_REF]: Map<EffectProperty | string, EffectSubscription> | null = null;
+
+  constructor(
+    container: Container | null,
+    fn: (...args: any[]) => T,
+    args: any[],
+    fnStr: string | null,
+    // We need a separate flag to know when the computation needs running because
+    // we need the old value to know if effects need running after computation
+    flags: SignalFlags = SignalFlags.INVALID | WrappedSignalFlags.UNWRAP
+  ) {
+    super(container, NEEDS_COMPUTATION);
+    this.$args$ = args;
+    this.$func$ = fn;
+    this.$funcStr$ = fnStr;
+    this.$flags$ = flags;
+  }
+
+  $invalidate$() {
+    this.$flags$ |= SignalFlags.INVALID;
+    this.$forceRunEffects$ = false;
+    // We should only call subscribers if the calculation actually changed.
+    // Therefore, we need to calculate the value now.
+    this.$container$?.$scheduler$(
+      ChoreType.RECOMPUTE_AND_SCHEDULE_EFFECTS,
+      this.$hostElement$,
+      this
+    );
+  }
+
+  /**
+   * Use this to force running subscribers, for example when the calculated value has mutated but
+   * remained the same object.
+   */
+  force() {
+    this.$flags$ |= SignalFlags.INVALID;
+    this.$forceRunEffects$ = false;
+    triggerEffects(this.$container$, this, this.$effects$);
+  }
+
+  get untrackedValue() {
+    const didChange = this.$computeIfNeeded$();
+    if (didChange) {
+      this.$forceRunEffects$ = didChange;
+    }
+    assertFalse(this.$untrackedValue$ === NEEDS_COMPUTATION, 'Invalid state');
+    return this.$untrackedValue$;
+  }
+
+  $computeIfNeeded$() {
+    if (!(this.$flags$ & SignalFlags.INVALID)) {
+      return false;
+    }
+    const untrackedValue = trackSignal(
+      () => this.$func$(...this.$args$),
+      this,
+      EffectProperty.VNODE,
+      this.$container$!
+    );
+    // TODO: we should remove invalid flag here
+    // this.$flags$ &= ~SignalFlags.INVALID;
+    const didChange = untrackedValue !== this.$untrackedValue$;
+    if (didChange) {
+      this.$untrackedValue$ = untrackedValue;
+    }
+    return didChange;
+  }
+  // Make this signal read-only
+  set value(_: any) {
+    throw qError(QError.wrappedReadOnly);
+  }
+  // Getters don't get inherited when overriding a setter
+  get value() {
+    return super.value;
+  }
+}
