+++ packages/qwik/src/core/signal/signal.ts
@@ -14,8 +14,7 @@
 import { isDomContainer } from '../client/dom-container';
 import { pad, qwikDebugToString } from '../debug';
 import type { OnRenderFn } from '../shared/component.public';
-import { assertDefined, assertFalse } from '../shared/error/assert';
-import { QError, qError } from '../shared/error/error';
+import { assertDefined } from '../shared/error/assert';
 import type { Props } from '../shared/jsx/jsx-runtime';
 import { type QRLInternal } from '../shared/qrl/qrl-class';
 import type { QRL } from '../shared/qrl/qrl.public';
@@ -29,7 +28,7 @@ import { TaskFlags, isTask } from '../use/use-task';
 import { NEEDS_COMPUTATION, _EFFECT_BACK_REF } from './flags';
 import { ComputedSignalImpl } from './impl/computed-signal-impl';
 import { SignalImpl } from './impl/signal-impl';
-import { type BackRef } from './signal-cleanup';
+import type { WrappedSignalImpl } from './impl/wrapped-signal-impl';
 import type { Signal } from './signal.public';
 import type { TargetType } from './store';
 import { SubscriptionData, type NodePropPayload } from './subscription-data';
@@ -37,8 +36,6 @@ import {
   EffectProperty,
   EffectSubscriptionProp,
   SignalFlags,
-  WrappedSignalFlags,
-  type AllSignalFlags,
   type ComputeQRL,
   type EffectSubscription,
 } from './types';
@@ -129,7 +126,7 @@ export const triggerEffects = (
           }
         }
 
-        (consumer as ComputedSignalImpl<unknown> | WrappedSignal<unknown>).$invalidate$();
+        (consumer as ComputedSignalImpl<unknown> | WrappedSignalImpl<unknown>).$invalidate$();
       } else if (property === EffectProperty.COMPONENT) {
         const host: HostElement = consumer as any;
         const qrl = container.getHostProp<QRLInternal<OnRenderFn<unknown>>>(host, OnRenderProp);
@@ -161,91 +158,6 @@ export const triggerEffects = (
 
   DEBUG && log('done scheduling');
 };
-
-export class WrappedSignal<T> extends SignalImpl<T> implements BackRef {
-  $args$: any[];
-  $func$: (...args: any[]) => T;
-  $funcStr$: string | null;
-
-  $flags$: AllSignalFlags;
-  $hostElement$: HostElement | null = null;
-  $forceRunEffects$: boolean = false;
-  [_EFFECT_BACK_REF]: Map<EffectProperty | string, EffectSubscription> | null = null;
-
-  constructor(
-    container: Container | null,
-    fn: (...args: any[]) => T,
-    args: any[],
-    fnStr: string | null,
-    // We need a separate flag to know when the computation needs running because
-    // we need the old value to know if effects need running after computation
-    flags: SignalFlags = SignalFlags.INVALID | WrappedSignalFlags.UNWRAP
-  ) {
-    super(container, NEEDS_COMPUTATION);
-    this.$args$ = args;
-    this.$func$ = fn;
-    this.$funcStr$ = fnStr;
-    this.$flags$ = flags;
-  }
-
-  $invalidate$() {
-    this.$flags$ |= SignalFlags.INVALID;
-    this.$forceRunEffects$ = false;
-    // We should only call subscribers if the calculation actually changed.
-    // Therefore, we need to calculate the value now.
-    this.$container$?.$scheduler$(
-      ChoreType.RECOMPUTE_AND_SCHEDULE_EFFECTS,
-      this.$hostElement$,
-      this
-    );
-  }
-
-  /**
-   * Use this to force running subscribers, for example when the calculated value has mutated but
-   * remained the same object.
-   */
-  force() {
-    this.$flags$ |= SignalFlags.INVALID;
-    this.$forceRunEffects$ = false;
-    triggerEffects(this.$container$, this, this.$effects$);
-  }
-
-  get untrackedValue() {
-    const didChange = this.$computeIfNeeded$();
-    if (didChange) {
-      this.$forceRunEffects$ = didChange;
-    }
-    assertFalse(this.$untrackedValue$ === NEEDS_COMPUTATION, 'Invalid state');
-    return this.$untrackedValue$;
-  }
-
-  $computeIfNeeded$() {
-    if (!(this.$flags$ & SignalFlags.INVALID)) {
-      return false;
-    }
-    const untrackedValue = trackSignal(
-      () => this.$func$(...this.$args$),
-      this,
-      EffectProperty.VNODE,
-      this.$container$!
-    );
-    // TODO: we should remove invalid flag here
-    // this.$flags$ &= ~SignalFlags.INVALID;
-    const didChange = untrackedValue !== this.$untrackedValue$;
-    if (didChange) {
-      this.$untrackedValue$ = untrackedValue;
-    }
-    return didChange;
-  }
-  // Make this signal read-only
-  set value(_: any) {
-    throw qError(QError.wrappedReadOnly);
-  }
-  // Getters don't get inherited when overriding a setter
-  get value() {
-    return super.value;
-  }
-}
 
 /** @public */
 export type SerializerArgObject<T, S> = {
